Improving some aspects in DAO's.

+++ Web Project/JuridicApp/src/java/co/edu/udea/juridicapp/persistence/entity/CommentPK.java
@@ -6,50 +6,56 @@ import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import javax.validation.constraints.NotNull;
 
+/**
+ *
+ * @author Miguel Ossa Ruiz
+ * @author Neiber Padierna P&eacute;rez
+ */
 @Embeddable()
 public class CommentPK implements Serializable {
 
     private static final long serialVersionUID = 6482725099022643200L;
     @Basic(optional = false)
     @Column(name = "id")
-    private Long id;
+    private long id;
     @Basic(optional = false)
     @NotNull()
     @Column(name = "oeuvre_id")
-    private Long oeuvreId;
+    private long oeuvreId;
 
     public CommentPK() {
         super();
     }
 
-    public CommentPK(Long id, Long oeuvreId) {
+    public CommentPK(long id, long oeuvreId) {
         this.id = id;
         this.oeuvreId = oeuvreId;
     }
 
-    public Long getId() {
+    public long getId() {
 
         return (this.id);
     }
 
-    public void setId(Long id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public Long getOeuvreId() {
+    public long getOeuvreId() {
 
         return (this.oeuvreId);
     }
 
-    public void setOeuvreId(Long oeuvreId) {
+    public void setOeuvreId(long oeuvreId) {
         this.oeuvreId = oeuvreId;
     }
 
     @Override()
     public int hashCode() {
         int hash = 0;
-        hash += (int) this.getId().longValue();
-        hash += (int) this.getOeuvreId().longValue();
+
+        hash += (int) this.getId();
+        hash += (int) this.getOeuvreId();
 
         return (hash);
     }
